move postgres test to location of query
